Add CHOP command to rxstrings
CHOP key count

Removes characters from a value to a String key.
If 'count' is positive or zero, it removes from the right of the string;
if 'count' is negative, it removes from the left of the string.
If |count| is greater than the string length, the value is set as an empty string.
It is an error if the value is not a String.
Integer Reply: the length of the string after the chop operation.